move makeKubeConfig out of pkg/config and condense
