+++ test/clients/openshift/login.go
@@ -1,15 +1,14 @@
 package openshift
 
 import (
-	"crypto/rsa"
 	"crypto/x509"
 	"crypto/x509/pkix"
 	"fmt"
-	"strings"
 
 	v1 "k8s.io/client-go/tools/clientcmd/api/v1"
 
 	internalapi "github.com/openshift/openshift-azure/pkg/api"
+	"github.com/openshift/openshift-azure/pkg/util/kubeconfig"
 	azuretls "github.com/openshift/openshift-azure/pkg/util/tls"
 )
 
@@ -40,58 +39,5 @@ func login(username string, cs *internalapi.OpenShiftManagedCluster) (*v1.Config
 		return nil, err
 	}
 
-	return makeKubeConfig(key, cert, cs.Config.Certificates.Ca.Cert, cs.Properties.FQDN, username, "default")
-}
-
-func makeKubeConfig(clientKey *rsa.PrivateKey, clientCert, caCert *x509.Certificate, endpoint, username, namespace string) (*v1.Config, error) {
-	clustername := strings.Replace(endpoint, ".", "-", -1)
-	authinfoname := username + "/" + clustername
-	contextname := namespace + "/" + clustername + "/" + username
-
-	caCertBytes, err := azuretls.CertAsBytes(caCert)
-	if err != nil {
-		return nil, err
-	}
-	clientCertBytes, err := azuretls.CertAsBytes(clientCert)
-	if err != nil {
-		return nil, err
-	}
-	clientKeyBytes, err := azuretls.PrivateKeyAsBytes(clientKey)
-	if err != nil {
-		return nil, err
-	}
-
-	return &v1.Config{
-		APIVersion: "v1",
-		Kind:       "Config",
-		Clusters: []v1.NamedCluster{
-			{
-				Name: clustername,
-				Cluster: v1.Cluster{
-					Server:                   "https://" + endpoint,
-					CertificateAuthorityData: caCertBytes,
-				},
-			},
-		},
-		AuthInfos: []v1.NamedAuthInfo{
-			{
-				Name: authinfoname,
-				AuthInfo: v1.AuthInfo{
-					ClientCertificateData: clientCertBytes,
-					ClientKeyData:         clientKeyBytes,
-				},
-			},
-		},
-		Contexts: []v1.NamedContext{
-			{
-				Name: contextname,
-				Context: v1.Context{
-					Cluster:   clustername,
-					Namespace: namespace,
-					AuthInfo:  authinfoname,
-				},
-			},
-		},
-		CurrentContext: contextname,
-	}, nil
+	return kubeconfig.Make(key, cert, cs.Config.Certificates.Ca.Cert, cs.Properties.FQDN, username, "default")
 }
